menyelesaikan register

+++ admin/daftar-user.php
@@ -1,11 +1,15 @@
 <?php
 
-
 session_start();
 
 if (!isset($_SESSION["login"])) {
-    header("Location: login");
+    header("Location: ../login");
+    return false;
+} elseif ($_SESSION["status"] != "admin") {
+    header("Location: index");
+    return false;
 }
+
 
 $username = $_SESSION["login"];
 
@@ -90,12 +94,8 @@ if ($_SESSION["status"] === "admin") {
                                         <span class="ms-1" id="sebagai"><?= $result["status"]; ?></span>
                                     </div>
                                 </a>
-                                <ul class="dropdown-menu dropdown-menu-light text-small shadow">
-                                    <li><a class="dropdown-item" href="#">Profile</a></li>
-                                    <li>
-                                        <hr class="dropdown-divider">
-                                    </li>
-                                    <li><a class="dropdown-item" href="logout">Sign out</a></li>
+                                <ul class="dropdown-menu dropdown-menu-light text-small border-0 shadow-sm">
+                                    <li><a class="dropdown-item text-danger" href="logout"> <i class="fa fa-sign-out"></i> Keluar</a></li>
                                 </ul>
                             </div>
                         </li>
@@ -191,10 +191,10 @@ if ($_SESSION["status"] === "admin") {
                                 <td><?= $crs["username"]; ?></td>
                                 <td><?= $crs["nomor_hp"]; ?></td>
                                 <td>
-                                    <a href="hapus-user?id_user=<?= $crs["id_user"] ?>" class="btn btn-outline-danger btn-sm" title="Hapus <?= $crs['nama_user'] ?> ?">
+                                    <a href="hapus-user?id_user=<?= $crs["id_user"] ?>" class="btn btn-outline-danger btn-sm" title="Hapus <?= $crs['nama_user'] ?> ?" onclick=" return confirm('Yakin Ingin Menghapus <?= $crs["nama_user"] ?>  ?')">
                                         <i class="fa fa-trash"></i>
                                     </a>
-                                    <a href="hapus-user?id_user=<?= $crs["id_user"] ?>" class="btn btn-outline-success btn-sm" title="Hapus <?= $crs['nama_user'] ?> ?">
+                                    <a href="akun?id_user=<?= $crs["id_user"] ?>" class="btn btn-outline-success btn-sm" title="Hapus <?= $crs['nama_user'] ?> ?">
                                         <i class="fa fa-eye"></i>
                                     </a>
                                 </td>
